signin and reg extend baseview; register button in SignInView changes the view to RegistrationView
