+++ src/app/BandHeroApp.java
@@ -6,6 +6,8 @@ import views.SignInView;
 import controllers.*;
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
 /**
  * Created by costin on 11/20/2015.
@@ -24,7 +26,7 @@ public class BandHeroApp {
         mainFrame = new JFrame("Band Hero");
         mainFrame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
         mainFrame.setPreferredSize(new Dimension(500, 400));
-        loadRegistrationView();
+        loadSignInView();
 
         mainFrame.pack();
         mainFrame.setVisible(true);
@@ -32,15 +34,26 @@ public class BandHeroApp {
 
     private void loadRegistrationView() {
         RegistrationView registrationView = new RegistrationView();
-        mainFrame.add(registrationView);
+        mainFrame.setContentPane(registrationView);
+        mainFrame.revalidate();
     }
 
     private void loadSignInView() {
         //Create SignInView and add it to the app Frame
         SignInView signInView = new SignInView();
+
+        signInView.addRegisterActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                BandHeroApp.getInstance().loadRegistrationView();
+            }
+        });
+
+
         mainFrame.add(signInView);
         //Add Sign In Action Listener from Controller.
         SignInController.getInstance().createSignInActionListener(signInView);
+
         try { UserDB.getUserDB().addUser("Costin", "12345", "Musician"); } catch (Exception e) {}
     }
